additional extra file copys without exist checks
see #3

+++ METRIC/prepare_metric_env.py
@@ -32,7 +32,8 @@ def copyfile(path, dest_dir, workspace = ""):
         return destination.replace(workspace + "\\", "")
 
     else:
-        raise IOError("{0} is an invalid filepath!".format(path))
+        print("{0} is an invalid filepath!".format(path))
+        return None
 
 
 def clip_inputs(raster, shapefile, outpath):
@@ -144,10 +145,11 @@ def prepare_metric_env(workspace, landsat_band2, landsat_band3, landsat_band4, l
 
 
     # move the DEM and associated files
-    copyfile(dem_path + ".ovr", dem_dir, workspace)
-    copyfile(dem_path + ".aux.xml", dem_dir, workspace)
-    copyfile(dem_path + ".xml", dem_dir, workspace)
-    copyfile(dem_path.replace(".tif", ".tfw"), dem_dir, workspace)
+    for demfile in [dem_path + ext for ext in [".ovr", ".aux.xml", ".xml"]]:
+        if os.path.exists(demfile):
+            copyfile(dem_path + ".ovr", dem_dir, workspace)
+    if os.path.exists(dem_path.replace(".tif", ".tfw")):
+        copyfile(dem_path.replace(".tif", ".tfw"), dem_dir, workspace)
 
     if clip_extent is None:
         dem_path = copyfile(dem_path, dem_dir, workspace)
